// apps/frontend/src/infrastructure/api/types/get-shopping-event-by-id200-products-item.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Grocery Store API
 * API de gerenciamento de compras
 * OpenAPI spec version: 1.0.0
 */

export type GetShoppingEventById200ProductsItem = {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  id: string;
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  canonicalProductId: string;
  name: string;
  amount: number;
  price: number;
  wholesaleMinAmount?: number;
  wholesalePrice?: number;
  totalRetailPrice: number;
  totalWholesalePrice: number;
  totalDifference?: number;
  addedAt: string;
};